pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "db-agent-chomi"
version = "0.1.0"
description = "Database agent project"
requires-python = ">=3.12"
dependencies = [
    "langchain",
    "python-dotenv",
    "langchain-openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
